+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
-- [...]
+- **[BREAKING CHANGE]** Migrate `DatePicker` to ReactDayPicker instead of React Dates, parts of the API of the component has been changed, and it no longer supports any language by default
+- **[BREAKING CHANGE]** Remove Moment dependency
 
 # v1.2.0 (20/03/2019)